Syntax errors fixed.
